aprs-legal-assistant/backend/requirements.txt
fastapi
uvicorn
python-dotenv
azure-cognitiveservices-speech
google-generativeai
google-cloud-translate
httpx
requests
python-multipart
pydantic
